refactor: rename schema to pb
