+++ Makefile
@@ -13,4 +13,4 @@ generate:
     --go_opt=paths=source_relative \
     --go-grpc_out=. \
     --go-grpc_opt=paths=source_relative \
-    grpc/schema/pool.proto
+    grpc/pb/pool.proto
